Исправляет ошибки тестирования для выделенного компонента секции комментариев
